test(parser): finish moving tests to their correct modules
