+++ crates/aiken-lang/src/parser/expr/bytearray.rs
@@ -12,3 +12,23 @@ pub fn parser() -> impl Parser<Token, UntypedExpr, Error = ParseError> {
         preferred_format,
     })
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::assert_expr;
+
+    #[test]
+    fn bytearray_basic() {
+        assert_expr!("#[0, 170, 255]");
+    }
+
+    #[test]
+    fn bytearray_base16() {
+        assert_expr!("#\"00aaff\"");
+    }
+
+    #[test]
+    fn bytearray_utf8_encoded() {
+        assert_expr!("\"aiken\"");
+    }
+}
