+++ crates/aiken-lang/src/tests/mod.rs
@@ -1,4 +1,3 @@
 mod check;
 mod format;
 mod lexer;
-mod parser;
